Move the config COTOAMI_SIGNUP_ENABLED to EmailAuthController
