+++ config/config.exs
@@ -22,6 +22,12 @@ config :logger, :console,
   format: "$time $metadata[$level] $message\n",
   metadata: [:request_id]
 
+# EmailAuth
+config :cotoami, CotoamiWeb.EmailAuthController,
+  signup_enabled:
+  (System.get_env("COTOAMI_SIGNUP_ENABLED") || "true")
+  |> String.to_existing_atom()
+
 # OAuth2
 config :oauth2, debug: true
 
@@ -34,10 +40,7 @@ config :cotoami, CotoamiWeb.OAuth2.Google,
 config :cotoami, Cotoami.AmishiService,
   owner_emails:
     (System.get_env("COTOAMI_OWNER_EMAILS") || "")
-    |> String.split(",", trim: true),
-  signup_enabled:
-    (System.get_env("COTOAMI_SIGNUP_ENABLED") || "true")
-    |> String.to_existing_atom()
+    |> String.split(",", trim: true)
 
 # Redis
 case System.get_env("REDIS_URL") do
