+++ lib/cotoami/services/amishi_service.ex
@@ -12,18 +12,6 @@ defmodule Cotoami.AmishiService do
     |> Keyword.get(:owner_emails)
   end
 
-  def signup_enabled do
-    :cotoami
-    |> Application.get_env(__MODULE__, [])
-    |> Keyword.get(:signup_enabled)
-  end
-
-  def is_allowed_to_signin?(email) do
-    signup_enabled() 
-      || email in owner_emails() 
-      || get_by_email(email)
-  end
-
   def get(id) do
     Amishi
     |> Repo.get(id)
@@ -31,7 +19,9 @@ defmodule Cotoami.AmishiService do
   end
 
   def get_by_email(email) do
-    Repo.get_by(Amishi, email: email)
+    Amishi
+    |> Repo.get_by(email: email)
+    |> append_owner_flag()
   end
 
   def insert_or_update_by_email!(email) do
